moved Google recogniser into own package, notes for VAD
