+++ src/main/java/com/darkprograms/speech/recognizer/google/RecognizerChunked.java
@@ -1,4 +1,4 @@
-package com.darkprograms.speech.recognizer;
+package com.darkprograms.speech.recognizer.google;
 
 
 import java.io.BufferedReader;
